pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragcompliance"
version = "0.1.0"
description = "Audit trail middleware for RAG pipelines in regulated industries"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Daksh Trehan" }]
keywords = ["rag", "langchain", "llamaindex", "compliance", "audit", "llm", "fintech", "healthtech"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain-core>=0.2.0",
    "stripe>=8.0.0",
]

[project.optional-dependencies]
supabase = ["supabase>=2.0.0"]
dashboard = ["fastapi>=0.110.0", "uvicorn>=0.29.0"]
llamaindex = ["llama-index-core>=0.10.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "langchain-core>=0.2.0",
    "ruff>=0.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/dakshtrehan/ragcompliance"
Repository = "https://github.com/dakshtrehan/ragcompliance"
Issues = "https://github.com/dakshtrehan/ragcompliance/issues"

[tool.hatch.build.targets.wheel]
packages = ["ragcompliance"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
